pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mediawhisperer"
version = "0.1.0"
description = "An interactive tool for processing and conversing with media content"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-multipart>=0.0.6",
    "python-jose>=3.3.0",
    "passlib[bcrypt]~=1.7.4",
    "bcrypt<4.0.0",
    "PyPDF2>=3.0.1",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "transformers>=4.35.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "ijson>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
]

[tool.setuptools]
packages = ["apps.api"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
